move constants to common folder, but when load page, the getServerFeature is still broken
