+++ models/interfaces.ts
@@ -7,8 +7,8 @@ import { Direction } from '@elastic/eui';
 import { WebhookMethodType } from '../public/pages/Channels/types';
 import {
   CHANNEL_TYPE,
-  ENCRYPTION_TYPE,
-} from '../public/utils/constants';
+} from '../common/constants';
+import {ENCRYPTION_TYPE} from '../public/utils/constants'
 
 export interface ChannelStatus {
   config_id: string;
