+++ server/routes/configRoutes.ts
@@ -11,7 +11,7 @@ import {
 import { NODE_API } from '../../common';
 import { joinRequestParams } from '../utils/helper';
 import _ from 'lodash';
-import { CHANNEL_TYPE } from '../../public/utils/constants';
+import { CHANNEL_TYPE } from '../../common/constants';
 
 export function configRoutes(router: IRouter) {
   router.get(
@@ -216,7 +216,7 @@ export function configRoutes(router: IRouter) {
           keyof typeof CHANNEL_TYPE
         >;
         const channelTypes: Partial<typeof CHANNEL_TYPE> = {};
-        
+
         for (let channel of config_type_list) {
           if (CHANNEL_TYPE[channel]) {
             channelTypes[channel] = CHANNEL_TYPE[channel]
@@ -229,6 +229,7 @@ export function configRoutes(router: IRouter) {
           tooltipSupport:
             _.get(response, ['plugin_features', 'tooltip_support']) === 'true',
         };
+        console.log("availableFeature", availableFeature.availableChannels)
         return response.ok({ body: availableFeature });
       } catch (error) {
         return response.custom({
